Initial working code. Needs cleaning up.

+++ utils/stack_utils.c
@@ -6,17 +6,18 @@
 /*   By: mahmoud <mahmoud@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/30 14:35:32 by mabdelsa          #+#    #+#             */
-/*   Updated: 2023/11/19 20:33:54 by mahmoud          ###   ########.fr       */
+/*   Updated: 2023/11/25 17:39:08 by mahmoud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-int	ft_stack_lstsize(t_stack *lst)
+size_t	ft_stack_lstsize(t_stack *lst)
 {
 	size_t	i;
 
 	i = 0;
+	// ft_printf("in size!!!");
 	while (lst)
 	{
 		i++;
